# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NutriCart API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface HealthProfile {
  id: number;
  userId: string;
  allergies: string[];
  conditions: string[];
  preferredPharmacies: string[];
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface HealthProfileUpdate {
  allergies?: string[];
  conditions?: string[];
  preferredPharmacies?: string[];
  /** @nullable */
  notes?: string | null;
}

export type MedicationType = typeof MedicationType[keyof typeof MedicationType];


export const MedicationType = {
  prescription: 'prescription',
  supplement: 'supplement',
  otc: 'otc',
} as const;

export interface Medication {
  id: number;
  userId: string;
  name: string;
  /** @nullable */
  genericName?: string | null;
  type: MedicationType;
  dosage: string;
  frequency: string;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  isActive: boolean;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  prescribedBy?: string | null;
  /** @nullable */
  pharmacy?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type MedicationInputType = typeof MedicationInputType[keyof typeof MedicationInputType];


export const MedicationInputType = {
  prescription: 'prescription',
  supplement: 'supplement',
  otc: 'otc',
} as const;

export interface MedicationInput {
  /** @minLength 1 */
  name: string;
  genericName?: string;
  type: MedicationInputType;
  /** @minLength 1 */
  dosage: string;
  /** @minLength 1 */
  frequency: string;
  startDate: string;
  endDate?: string;
  notes?: string;
  prescribedBy?: string;
  pharmacy?: string;
}

export type MedicationUpdateType = typeof MedicationUpdateType[keyof typeof MedicationUpdateType];


export const MedicationUpdateType = {
  prescription: 'prescription',
  supplement: 'supplement',
  otc: 'otc',
} as const;

export interface MedicationUpdate {
  /** @minLength 1 */
  name?: string;
  genericName?: string;
  type?: MedicationUpdateType;
  dosage?: string;
  frequency?: string;
  startDate?: string;
  /** @nullable */
  endDate?: string | null;
  isActive?: boolean;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  prescribedBy?: string | null;
  /** @nullable */
  pharmacy?: string | null;
}

export type InteractionFlagSeverity = typeof InteractionFlagSeverity[keyof typeof InteractionFlagSeverity];


export const InteractionFlagSeverity = {
  mild: 'mild',
  moderate: 'moderate',
  severe: 'severe',
} as const;

export interface InteractionFlag {
  severity: InteractionFlagSeverity;
  description: string;
  medications: string[];
  /** @nullable */
  recommendation?: string | null;
}

export interface InteractionReport {
  checkedAt: string;
  flags: InteractionFlag[];
  checkedMedications: string[];
  isClear?: boolean;
}

export interface PharmacyPrice {
  pharmacyName: string;
  price: number;
  unit: string;
  inStock: boolean;
  deliveryAvailable?: boolean;
  /** @nullable */
  couponCode?: string | null;
  /** @nullable */
  couponPrice?: number | null;
  /** @nullable */
  distance?: string | null;
}

export interface PriceComparisonResult {
  medicationName: string;
  dosage: string;
  prices: PharmacyPrice[];
  bestPrice: number;
  avgPrice: number;
  /** @nullable */
  bestPharmacy?: string | null;
}

export interface SavingsEntry {
  id: number;
  userId: string;
  medicationName: string;
  listPrice: number;
  paidPrice: number;
  savedAmount: number;
  pharmacyName: string;
  purchaseDate: string;
  /** @nullable */
  couponUsed?: string | null;
  createdAt: string;
}

export interface SavingsInput {
  /** @minLength 1 */
  medicationName: string;
  listPrice: number;
  paidPrice: number;
  /** @minLength 1 */
  pharmacyName: string;
  purchaseDate: string;
  couponUsed?: string;
}

export interface MonthlySaving {
  month: string;
  saved: number;
  spent: number;
}

export interface SavingsSummary {
  totalSaved: number;
  totalSpent: number;
  savingsCount: number;
  avgSavingPerPurchase: number;
  /** @nullable */
  topPharmacy: string | null;
  monthlySavings: MonthlySaving[];
}

export interface Reminder {
  id: number;
  userId: string;
  medicationName: string;
  refillDate: string;
  daysSupply: number;
  /** @nullable */
  currentSupplyDays?: number | null;
  isDismissed: boolean;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ReminderInput {
  /** @minLength 1 */
  medicationName: string;
  refillDate: string;
  /** @minimum 1 */
  daysSupply: number;
  /** @minimum 0 */
  currentSupplyDays?: number;
  notes?: string;
}

export interface ReminderUpdate {
  medicationName?: string;
  refillDate?: string;
  /** @minimum 1 */
  daysSupply?: number;
  /**
     * @minimum 0
     * @nullable
     */
  currentSupplyDays?: number | null;
  isDismissed?: boolean;
  /** @nullable */
  notes?: string | null;
}

export interface DashboardSummary {
  activeMedicationsCount: number;
  upcomingRemindersCount: number;
  totalSavedAllTime: number;
  interactionFlagsCount: number;
  recentSavings: SavingsEntry[];
  upcomingReminders: Reminder[];
}

export type RecommendationType = typeof RecommendationType[keyof typeof RecommendationType];


export const RecommendationType = {
  generic_alternative: 'generic_alternative',
  otc_option: 'otc_option',
  savings_tip: 'savings_tip',
  interaction_warning: 'interaction_warning',
  refill_reminder: 'refill_reminder',
} as const;

export type RecommendationRelevance = typeof RecommendationRelevance[keyof typeof RecommendationRelevance];


export const RecommendationRelevance = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface Recommendation {
  id: string;
  title: string;
  description: string;
  type: RecommendationType;
  relevance: RecommendationRelevance;
  /** @nullable */
  medicationName?: string | null;
  /** @nullable */
  estimatedSavings?: number | null;
  /** @nullable */
  actionLabel?: string | null;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};

export type LogoutBrowserSessionParams = {
returnTo?: string;
};

export type SearchMedicationPricesParams = {
name: string;
};