# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NutriCart API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get the current user's health profile
      responses:
        "200":
          description: User health profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Update the current user's health profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HealthProfileUpdate"
      responses:
        "200":
          description: Updated health profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # Medications
  /medications:
    get:
      operationId: listMedications
      tags: [medications]
      summary: List all medications for the current user
      responses:
        "200":
          description: List of medications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Medication"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createMedication
      tags: [medications]
      summary: Add a new medication to the user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicationInput"
      responses:
        "201":
          description: Medication created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /medications/{id}:
    get:
      operationId: getMedication
      tags: [medications]
      summary: Get a single medication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Medication found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateMedication
      tags: [medications]
      summary: Update a medication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicationUpdate"
      responses:
        "200":
          description: Medication updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteMedication
      tags: [medications]
      summary: Remove a medication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # Interactions
...
  /prices/search:
    get:
      operationId: searchMedicationPrices
      tags: [prices]
      summary: Search pharmacy prices for a medication by name
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Price comparison results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceComparisonResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /prices/compare/{medicationId}:
    get:
      operationId: getMedicationPrices
      tags: [prices]
      summary: Get price comparison for one of the user's medications
      parameters:
        - name: medicationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Price comparison results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceComparisonResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # Savings
...
  /reminders:
    get:
      operationId: listReminders
      tags: [reminders]
      summary: List all refill reminders for the current user
      responses:
        "200":
          description: List of reminders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reminder"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createReminder
      tags: [reminders]
      summary: Create a refill reminder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderInput"
      responses:
        "201":
          description: Reminder created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /reminders/{id}:
    patch:
      operationId: updateReminder
      tags: [reminders]
      summary: Update a reminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderUpdate"
      responses:
        "200":
          description: Reminder updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteReminder
      tags: [reminders]
      summary: Delete a reminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # Dashboard
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # Profile schemas
    HealthProfile:
      type: object
      required: [id, userId, allergies, conditions, preferredPharmacies, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        allergies:
          type: array
          items:
            type: string
        conditions:
          type: array
          items:
            type: string
        preferredPharmacies:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    HealthProfileUpdate:
      type: object
      properties:
        allergies:
          type: array
          items:
            type: string
        conditions:
          type: array
          items:
            type: string
        preferredPharmacies:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]

    # Medication schemas
    Medication:
      type: object
      required: [id, userId, name, type, dosage, frequency, startDate, isActive, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        genericName:
          type: ["string", "null"]
        type:
          type: string
          enum: [prescription, supplement, otc]
        dosage:
          type: string
        frequency:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        isActive:
          type: boolean
        notes:
          type: ["string", "null"]
        prescribedBy:
          type: ["string", "null"]
        pharmacy:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    MedicationInput:
      type: object
      required: [name, type, dosage, frequency, startDate]
      properties:
        name:
          type: string
          minLength: 1
        genericName:
          type: string
        type:
          type: string
          enum: [prescription, supplement, otc]
        dosage:
          type: string
          minLength: 1
        frequency:
          type: string
          minLength: 1
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        notes:
          type: string
        prescribedBy:
          type: string
        pharmacy:
          type: string
    MedicationUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        genericName:
          type: string
        type:
          type: string
          enum: [prescription, supplement, otc]
        dosage:
          type: string
        frequency:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        isActive:
          type: boolean
        notes:
          type: ["string", "null"]
        prescribedBy:
          type: ["string", "null"]
        pharmacy:
          type: ["string", "null"]

    # Interaction schemas
...
    PharmacyPrice:
      type: object
      required: [pharmacyName, price, unit, inStock]
      properties:
        pharmacyName:
          type: string
        price:
          type: number
        unit:
          type: string
        inStock:
          type: boolean
        deliveryAvailable:
          type: boolean
        couponCode:
          type: ["string", "null"]
        couponPrice:
          type: ["number", "null"]
        distance:
          type: ["string", "null"]
    PriceComparisonResult:
      type: object
      required: [medicationName, dosage, prices, bestPrice, avgPrice]
      properties:
        medicationName:
          type: string
        dosage:
          type: string
        prices:
          type: array
          items:
            $ref: "#/components/schemas/PharmacyPrice"
        bestPrice:
          type: number
        avgPrice:
          type: number
        bestPharmacy:
          type: ["string", "null"]

    # Savings schemas
...
    Reminder:
      type: object
      required: [id, userId, medicationName, refillDate, daysSupply, isDismissed, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        medicationName:
          type: string
        refillDate:
          type: string
          format: date
        daysSupply:
          type: integer
        currentSupplyDays:
          type: ["integer", "null"]
        isDismissed:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ReminderInput:
      type: object
      required: [medicationName, refillDate, daysSupply]
      properties:
        medicationName:
          type: string
          minLength: 1
        refillDate:
          type: string
          format: date
        daysSupply:
          type: integer
          minimum: 1
        currentSupplyDays:
          type: integer
          minimum: 0
        notes:
          type: string
    ReminderUpdate:
      type: object
      properties:
        medicationName:
          type: string
        refillDate:
          type: string
          format: date
        daysSupply:
          type: integer
          minimum: 1
        currentSupplyDays:
          type: ["integer", "null"]
          minimum: 0
        isDismissed:
          type: boolean
        notes:
          type: ["string", "null"]

    # Dashboard schema
...