# lib/api-zod/src/generated/types/medicationUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NutriCart API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MedicationUpdateType } from './medicationUpdateType';

export interface MedicationUpdate {
  /** @minLength 1 */
  name?: string;
  genericName?: string;
  type?: MedicationUpdateType;
  dosage?: string;
  frequency?: string;
  startDate?: Date;
  /** @nullable */
  endDate?: Date | null;
  isActive?: boolean;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  prescribedBy?: string | null;
  /** @nullable */
  pharmacy?: string | null;
}