# lib/api-zod/src/generated/types/priceComparisonResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NutriCart API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PharmacyPrice } from './pharmacyPrice';

export interface PriceComparisonResult {
  medicationName: string;
  dosage: string;
  prices: PharmacyPrice[];
  bestPrice: number;
  avgPrice: number;
  /** @nullable */
  bestPharmacy?: string | null;
}